server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
supabase==2.0.2
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6
pandas==2.1.4
nltk==3.8.1
google-generativeai==0.3.2
